Remove the dependency on prime.
Significant changes in lib/{delegation,events}.js with respect to the
datastructures.

Everything passes jshint now, with the .jshintrc file in this repo.

Didn't use the Array.prototype in list.js anymore if it would exist. The simple
JS version is probably equally fast and doesn't implement special cases the
native version implements, because it simply assumes it's an elements
collection.