.unlink() now calls $.clean functions on the element. Add $.gc
This way external code can clean up when an element is unlinked.
For example when you have a Drag class, the drag instance becomes useless when
you unlink/remove the element. $.clean.drag, for example, can detach event
handlers so it can be cleaned up nicely.

Another thing is $.gc, which is a function that checks for elements in the
instances object, that are not in the DOM anymore. To check this the .attached
method was added.

At a $.gc call, not all detached elements should be unlinked. To protect those
the .protect and inverse .unprotect are added.